requirements.txt
streamlit
pandas
beautifulsoup4
requests
lxml
fpdf2